reorganize session and device
