+++ tiny_cnn/core/core.h
@@ -24,6 +24,5 @@
 */
 #pragma once
 
-#include "tiny_cnn/core/session.h"
-
 #include "tiny_cnn/core/framework/device_base.h"
+#include "tiny_cnn/core/framework/session.h"
