--- tiny_cnn/core/device.h
@@ -1,79 +1,0 @@
-/*
-    Copyright (c) 2016, Taiga Nomi, Edgar Riba
-    All rights reserved.
-
-    Redistribution and use in source and binary forms, with or without
-    modification, are permitted provided that the following conditions are met:
-    * Redistributions of source code must retain the above copyright
-    notice, this list of conditions and the following disclaimer.
-    * Redistributions in binary form must reproduce the above copyright
-    notice, this list of conditions and the following disclaimer in the
-    documentation and/or other materials provided with the distribution.
-    * Neither the name of the <organization> nor the
-    names of its contributors may be used to endorse or promote products
-    derived from this software without specific prior written permission.
-
-    THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY
-    EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
-    WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
-    DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY
-    DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
-    (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
-    LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
-    ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-    (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
-    SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
-*/
-#pragma once
-
-#include <vector>
-
-#include "tiny_cnn/core/backend.h"
-
-#ifdef CNN_HAVE_OPENCL
-#ifdef __APPLE__
-#include <OpenCL/opencl.h>
-#else
-#include <CL/cl.h>
-#endif
-#endif
-
-#ifdef CNN_USE_LIBDNN
-#include "libdnn.hpp"
-#endif
-
-namespace tiny_cnn {
-namespace core {
-
-class device {
- public:
-  explicit device(const int id, const int list_id,
-                  const backend_t backend_type=backend_t::libdnn) {
-#ifndef CNN_USE_LIBDNN
-    }
-#else
-    if (backend_type == backend_t::tiny_cnn ||
-        backend_type == backend_t::nnpack) {
-        dev_ptr_ = std::make_shared<greentea::device>(
-            id, list_id, greentea::Backend::BACKEND_CPU);
-    } else if (backend_type == backend_t::libdnn) {
-         dev_ptr_ = std::make_shared<greentea::device>(
-            id, list_id, greentea::Backend::BACKEND_OpenCL);
-    } else {
-        throw nn_error("Not supported backend type");
-    }
-  }
-
-    int get_id() const { return id_; }
-
- private:
-  std::shared_ptr<greentea::device> dev_ptr_;
-
-  // TODO(edgar): revise API and decide how to handle this
-  std::vector<std::shared_ptr<backend>> backends_;
-#endif
-};
-
-}  // namespace core
-}  // namespace tiny_cnn
-
